backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.4.0
pymongo==4.6.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.5.0
bcrypt==5.0.0
python-dotenv==1.0.1
email-validator==2.2.0
pytest==9.0.2
